Aula 31/08: Response Entity; Embeddeable/Embedded; Ordenação de tarefas por prioridade

+++ projeto-final/turma-java-julho-2023-noite/taskify-api/src/main/java/com/taskify/api/controller/UsuarioController.java
@@ -8,6 +8,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,12 +27,12 @@ import com.taskify.api.repository.UsuarioRepository;
 public class UsuarioController {
 
     @PostMapping
-    public Usuario cadastrarUsuario(@RequestBody Usuario usuario) {
-        return usuarioRepository.save(usuario);
+    public ResponseEntity<Usuario> cadastrarUsuario(@RequestBody Usuario usuario) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(usuarioRepository.save(usuario));
     }
 
     @PutMapping("/{id}")
-    public Usuario atualizarUsuario(@PathVariable("id") Long id, @RequestBody Usuario usuario) {
+    public ResponseEntity<Usuario> atualizarUsuario(@PathVariable("id") Long id, @RequestBody Usuario usuario) {
         Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
 
         if (usuarioExistente.isPresent()) {
@@ -42,26 +44,40 @@ public class UsuarioController {
             usuarioObj.setSenha(usuario.getSenha());
             usuarioObj.setGenero(usuario.getGenero());
 
-            return usuarioRepository.save(usuarioObj);
+            return ResponseEntity.status(HttpStatus.OK).body(usuarioRepository.save(usuarioObj));
         }
 
-        return null;
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @GetMapping
-    public Page<Usuario> listarUsuarios(
-        @PageableDefault(size = 2, page = 1, sort = "email", direction = Direction.DESC) Pageable paginacao) {
-        return usuarioRepository.findAll(paginacao);
+    public ResponseEntity<Page<Usuario>> listarUsuarios(
+            @PageableDefault(size = 2, page = 1, sort = "email", direction = Direction.DESC) Pageable paginacao) {
+        return ResponseEntity.status(HttpStatus.OK).body(usuarioRepository.findAll(paginacao));
     }
 
     @GetMapping("/{id}")
-    public Optional<Usuario> obterUsuarioPeloId(@PathVariable("id") Long id) {
-        return usuarioRepository.findById(id);
+    public ResponseEntity<Optional<Usuario>> obterUsuarioPeloId(@PathVariable("id") Long id) {
+        Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
+
+        if (usuarioExistente.isPresent()) {
+            return ResponseEntity.status(HttpStatus.OK).body(usuarioExistente);
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @DeleteMapping("/{id}")
-    public void deletarUsuarioPeloId(@PathVariable("id") Long id) {
+    public ResponseEntity<Void> deletarUsuarioPeloId(@PathVariable("id") Long id) {
+        Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
+
+        if (usuarioExistente.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
         usuarioRepository.deleteById(id);
+
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
     @GetMapping("/email/{email}")
@@ -74,7 +90,6 @@ public class UsuarioController {
         return usuarioRepository.findByNome(nome);
     }
 
-    // Ponto de injeção
     @Autowired
     private UsuarioRepository usuarioRepository;
